move marker transfer to _pytest.mark
